# refactor(protocol): faster appendHeaderValue