--- a/internal/bytesconv/bytesconv_timing_test.go
+++ b/internal/bytesconv/bytesconv_timing_test.go
@@ -17,7 +17,6 @@
 package bytesconv
 
 import (
-	"strings"
 	"testing"
 )
 
@@ -49,51 +48,6 @@ func BenchmarkValidHeaderFiledValueTableHertz(b *testing.B) {
 			_ = func() bool {
 				return ValidHeaderFieldValueTable[s] != 0
 			}()
-		}
-	}
-}
-
-func BenchmarkNewlineToSpace(b *testing.B) {
-	// Test all characters
-	allBytes := make([]byte, 0)
-	for i := 0; i < 256; i++ {
-		allBytes = append(allBytes, byte(i))
-	}
-	headerNewlineToSpace := strings.NewReplacer("\n", " ", "\r", " ")
-
-	for i := 0; i < b.N; i++ {
-		_ = headerNewlineToSpace.Replace(string(allBytes))
-	}
-}
-
-func BenchmarkNewlineToSpaceHertz01(b *testing.B) {
-	// Test all characters
-	allBytes := make([]byte, 0)
-	for i := 0; i < 256; i++ {
-		allBytes = append(allBytes, byte(i))
-	}
-
-	for i := 0; i < b.N; i++ {
-		filteredVal := make([]byte, 0, len(allBytes))
-		for i := 0; i < len(allBytes); i++ {
-			filteredVal = append(filteredVal, NewlineToSpaceTable[allBytes[i]])
-		}
-		_ = filteredVal
-	}
-}
-
-func BenchmarkNewlineToSpaceHertz02(b *testing.B) {
-	// Test all characters
-	allBytes := make([]byte, 0)
-	for i := 0; i < 256; i++ {
-		allBytes = append(allBytes, byte(i))
-	}
-
-	for i := 0; i < b.N; i++ {
-		filteredVal := make([]byte, len(allBytes))
-		copy(filteredVal, allBytes)
-		for ii := 0; ii < len(allBytes); ii++ {
-			filteredVal[ii] = NewlineToSpaceTable[filteredVal[ii]]
 		}
 	}
 }