… internal/bytesconv/bytesconv_table.go